split company_model tests to seperate files
