+++ __tests__/server.spec.js
@@ -1,11 +1,11 @@
 const request = require('supertest');
-const server = require('./server');
+const server = require('../api/server');
 const db = require('../data/dbConfig');
 
 describe('server.js', () => {
-	it('runs the tests', () => {
-		expect(true).toBe(true);
+  it('runs the tests', () => {
+    expect(true).toBe(true);
 
-		//add more tests here
-	});
+    //add more tests here
+  });
 });
